Refactor ProtectedRoute component and update import path
